reflow/values: fix digests for directories
Summary:
The digest computation for directory values did not include directory
keys (though they were sorted for this purpose!). The effect of this
is that two directories that were identical by content (but not
necessarily keys) would digest to the same value. This would then
later cause canonicalization to (wrongly) merge nodes, and discard
one of the alternatives.

This manifested itself in a bug where attempting to add a prefix to
directory entries would fail, e.g.:

	val xyz = dir("s3://grail-marius/fastqs1/")
	val Main = dirs.Make(map([("XXX"+name, f) | (name, f) <- map(xyz)]))

Reviewers: pgopal, escott

Reviewed By: pgopal, escott

Differential Revision: https://phabricator.grailbio.com/D7189

fbshipit-source-id: 3e7227e