reflow/runner: improve stealer allocation
Summary:
The work stealer queries the evaluator for the amount of queued up
work that's available. However, it uses the flow's minimum
requirements as its minimum. This has two problems:

1. It can create allocs that are too large for the job at hand.
2. The min resource requirements can be more than the total.

Retrieving the minimum was really just a proxy for making sure we had
enough space to fit at least one stealable task. This change
introduces a new datastructure, Requirements, which exposes this
directly.

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D7303

fbshipit-source-id: 9ba0f35