reflow/reflowlet: terminate on 10-minute idle time
Summary:
Currently, the Reflowlet only terminates when it is idle near a billing
boundary. Now that Amazon has per-second billing, this is obsolete.
This changes this behavior so that instances are terminated if they
have been idle for 10 minutes. This allows instances to be reused,
but also not consume too many extra billing seconds.

Reviewers: pgopal, razvanm

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D7334

fbshipit-source-id: 302fe32