reflow/syntax,types: fix type environment for recursive module instantiations
Summary:
Each module instantiation uses the session's Types environment, and
(*Module).Make is careful to push and then pop the environment.
However, when modules a second module instantiates yet another, it
uses the same type environment (which would need to be popped, and
restored again).

Instead, switch over to a linked environment like we have in
values.Env, which makes bugs like much less likely and also fixes
this one.

Reviewers: pgopal, sbagaria

Reviewed By: sbagaria

Differential Revision: https://phabricator.grailbio.com/D7400

fbshipit-source-id: 44db2c5