reflow: scramble digest when an OpVal carries an error
Summary:
Reflow's digest code currently does not digest error values. This is
OK from a caching perspective since errors are never stored, but it
means that when the graph is canonicalized (dynamically), a digest
for an empty value is indistinguishable from the digest of an error.
This thus invokes a race: the first empty value wins, and thus can
cause the evaluator to swallow errors.

Reviewers: sbagaria, pgopal

Reviewed By: sbagaria

Differential Revision: https://phabricator.grailbio.com/D7413

fbshipit-source-id: e482990