reflow/syntax: fix environment persistence for declarations
Summary:
The evaluator persists the environment before declarations to
maintain lexical scoping. However, in blocks and in module declaration
lists, the environment was not properly persisted. In the case where
an inner identifier (e.g., in a function's block expression) was
also defined as a previous declaration, this would cause the inner
declaration to be redefined to the latest one, and, during evaluation,
cause infinite recursion because the closure's environment refered
to itself.

Fix this by persisting the environment properly before each value
binding in declarations and blocks.

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D7438

fbshipit-source-id: e8f31b1