reflow/syntax: fix handling of delayed arguments in ExprMake, digest1
Summary:
D7567 introduced the ability to synthesize v1 Reflow modules from v0
scripts. In doing so, we had to permit module instantiation ("make")
to be delayed, because the v0 interpreter does not permit lazy
evaluation, and its arguments must be forced. We left out the fact
that (*Expr).k requires tvals, and that digest1 must implemented for
ExprMake.

Reviewers: sbagaria, pgopal

Reviewed By: sbagaria

Differential Revision: https://phabricator.grailbio.com/D7645

fbshipit-source-id: 9b8765b