reflow/ec2cluster: increase capacity check timeout; interpret timeouts correctly
Summary:
Ec2cluster relies on interpreting timeouts correctly (to maintain
instance capacity state). The code assumed that context errors would
be returned as-is, but the AWS Go SDK wraps it into an awserr with a
custom code. Interpret these correctly.

Since we're observing these timeouts (maybe transiently), also double it.

Reviewers: joshnewman

Reviewed By: joshnewman

Differential Revision: https://phabricator.grailbio.com/D8261

fbshipit-source-id: 0b1f0d2