reflow/cache: record last access time
Summary:
This change modifies an assoc's last access time every time it is
looked up.  This will allow us to implement a mark-and-sweep LRU
garbage collector: filter out the dynamodb table by access time.
These are now the roots of the object walk. Now construct a bloom
filter with the live objects, scan s3 and remove the ones not in
the filter.

Reviewers: razvanm, cnicolaou

Reviewed By: razvanm, cnicolaou

Differential Revision: https://phabricator.grailbio.com/D8484

fbshipit-source-id: cc7ea54