reflow: also use physical keys when storing cache entries; look up both
Summary:
This change adds the ability for flow nodes to define multiple cache
keys. All keys are stored and looked up. To simplify things, this
change also unifies the cache write path and cleans up a few other
things.

If needed, we can add read repair later. (This could be useful once
we begin defining cache keys that are not computable from the flow
nodes directly.)

Reviewers: pgopal, sbagaria

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D8779

fbshipit-source-id: f5d5a9a