reflow: perform cache read repair
Summary:
This change modifies Reflow's evaluator to write back missing cache
keys. That is, when there is a cache hit, we pick the first
non-missing key, and write it back to all the other keys. This allows
us to keep synonymous cache keys up to date at all times.

Possibly we should push this down to the cache in the future so that the
cache manages synonyms (and their representation).

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D8989

fbshipit-source-id: 373e71c