reflow/eval: use a stricter interpretation of empty values when recomputing empty values
Summary:
Currently, values are recomputed when they are directly empty (empty
map or list with only empty items). This doesn't consider the cases
where a single exec has multiple ouptuts (which are placed in lists),
and produces a combination of empty and nonempty values. This change
makes this check stricter, so that values that have any empty
subvalues are recomputed.

Reviewers: pgopal, sbagaria

Reviewed By: pgopal, sbagaria

Differential Revision: https://phabricator.grailbio.com/D9129

fbshipit-source-id: 0eb0249