reflow/local: use a fresh timer for each attempt to decode Docker's profile JSON
Summary:
We have seen (*dockerExec).profile hang attempting to drain a stopped
timer. This shouldn't happen according to the docs for time.Timer,
but nevertheless we have stack traces to support it.

Since we always wait this timer out, except for when we're done
profiling (through context cancellation), there's no real benefit in
reusing the timer.

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D9134

fbshipit-source-id: 11b0d29