reflow/repository: fix improper usage of HTTP HEAD
Summary:
We were abusing HTTP, and Go was having none of it. The remote
repository's Stat implementation used HTTP head to query file
metadata. Technically, HEAD requests are not supposed to return
reponse bodies; they are supposed to replicate only the metadata you
would get with a successful GET request. Go's http package enforces
this behavior, and dropped the server's body. This caused errors when
Stat was called. This was not problematic since the repositories are
clever enough to transfer only those objects that are needed, but it
caused inaccurate reporting of needed transfers during evaluation.

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D9370

fbshipit-source-id: 491a2b0