reflow/assoc: add larger timeouts and concurrency limits to assoc requests
Summary:
Some workflows can emit a large rate of assoc.Get requests,
especially when they are restarted, or evaluated in bottom-up mode.
This caused two operational issues: the cache timeout was not
generous enough, and, because AWS uses http/1.1, a lot of time was
spent establishing new connections, when reusing them would actually
be faster. This change adds "band-aid" to these two problems by (1)
reintroducing concurrency limits on assoc calls; (2) increasing the
over all cache lookup timeout to an almost absurd value.

The proper solution is left in a comment in this change.

Reviewers: afields, pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D9429

fbshipit-source-id: 8f83061