reflow: extensible resources; use requirements in cluster allocation
Summary:
This change sets the stage for adding further resource requirements
in Reflow. It changes reflow.Resources to be an arbitrary mapping of
labels to (floating point) values; disk, cpu, and mem are now keys in
this mapping.

We also change the cluster allocator to accept reflow.Requirements
directly instead of a min and max. This also means we no longer need
to use the sentinel value reflow.MaxResources to indicate "wide"
resource requests. These are now an explicit flag in the
requirements.

Reviewers: pgopal, sgross

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D9473

fbshipit-source-id: de4d2fc