reflow: use concurrency limits only for transfers; make them configurable
Summary:
Reflow limited transfers by amount of pending transfer. I'd be hard
pressed to tell you why this was a good strategy (especially since
node-to-node transfers are per-file single-streamed; and s3 transfers
anyway had infinite limits; maybe the theory was that transferring
lots of small files would be penalized too much in a concurrency
based policy) This change converts these limits to be concurrency
based only. This limit has a default of 20 (per node) and is
configurable with the key "transferlimit".

Reviewers: pgopal, sgross

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D9780

fbshipit-source-id: 25d5999