reflow/ec2cluster: a few scheduler tweaks, simplification
Summary:
This change adds better accounting for pending instances in the
ec2cluster scheduler, by following the same original packing order.
This prevents overallocation. With this change we can also safely
remove gating of batch runs.

We also simplify the many variants on comparisons on
reflow.Resources. It's now wittled down to Resources.Available; we
re-phrase other uses in terms of this.

Not done: we really need to restructure this code so that it's
testable.

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D10167

fbshipit-source-id: 4c483f2