reflow: add status reporting, render to terminal
Summary:
This change adds a system for status reporting to Reflow. A status
comprises zero or more groups, which in turn comprise zero or more
tasks. Each task has a beginning and end, and its status can be
updated at any time.

If Reflow is running in a capable terminal, the latest status is
displayed and updated at a regular interval. The display code goes
through some lengths to have good UX for this. When running in a dumb
terminal or no terminal at all, statuses are rendered when they are
changed, with a minimum interval between updates. The display code
also goes through some lengths to ensure that regular log messages
(and any writes to stdout or stderr) are interleaved properly with
status updates.

We implement status updates for:
- evaluation: each active exec, intern, or extern are represented as
  tasks, with updates on state changes;
- ec2cluster: each instance launch is represented as a task, and
  it is updated regularly while allocating new instances;
- batch: each run inside of a batch is represented as a task, updated
  with its state
- transfers: each transfer is represented as a task, displaying waiting,
  pending, and completed transfers (files, size)

Anecdotally, this vastly improves the UX for understanding what
Reflow is doing when.

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D10145

fbshipit-source-id: 9712f03