reflow/syntax: support rich(er) key types
Summary:
We now admit string, int, float, bool, file. This could in theory be
expanded to any type that is a composite of those basic types, but
this is probably not generally useful.

Doing so requires us to move away from using the standard Go map
as an underlying data structure. Instead we use a standard Go map as
a hash table (keyed by key digest) to a list of map entries.

Reviewers: ysaito, pgopal

Reviewed By: ysaito

Differential Revision: https://phabricator.grailbio.com/D9996

fbshipit-source-id: 3ca52b4