reflow/internal/wg: make background waitgroups compositional; wait for cache writes in stealers
Summary:
Previously we relied on the stealer polling timeout to terminate
cache writes. This revision changes the behavior so that we
explicitly control cache write timeouts; and we give it a generous
one hour to complete cache writes. We have some cases where the
worked output is very valuable, and we were discarding it because we
didn't give the cache writes enough time.

This revision also makes the wgs compositional by using channels as
the notification mechanism. This way we can print sensisible log
messages when we're waiting for cache writes.

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D10165

fbshipit-source-id: 222bc57