reflow/repository: single-flight transfers in repository.Manager; roll up stats
Summary:
Transfer requests currently are single-flighted at the repository
level. While this correctly avoids duplicate concurrent transfers,
the higher layers (the transfer manager) doesn't know the difference
between a shared transfer and a non-shared transfer. This can
sometimes make that transfer stats baffling: for example, huge
amounts of data transfer can be reported because a shared reference
is transferered with every exec; in reality, that reference is
transferred exactly once.

This change addresses this problem and also adds rollups to the
transfer status display, to make them more comprehensible. We do this
by introducing single-flighting of requests in the repository
manager, and also by keeping track of aggregate transfer statistics
as well as transfer statistics per task. A transfer task is defined
as all transfers rolled up by (src, dst). This means that users will
see a status line for each (src, dst) pair, so long as there is an
active transfer between them, as well as the aggregate transfer
statistics.

TBA: To Be Audited.

Reviewers: pgopal

Differential Revision: https://phabricator.grailbio.com/D10433

fbshipit-source-id: cc1c4ee