reflow: model width explicitly in requirements; account for this when allocating
Summary:
Currently, resource requirements are a simple range and a "wide"
flag. This accomodate for two uses cases: (1) graphs with dynamic
widths (e.g., OpMap, and of course OpK, which can do anything) so
that we can account for this when performing resource allocation
(these graphs are likely to need some multiple of the minimum
resource requirements); (2) represent work-stealing requirements
where we have a concrete set (usually equally-sized) tasks. The two
interacted in sometimes strange ways, and required us to contort
Requirements to do our bidding (e.g., using MaxResources to represent
an unknown upper bound). It also caused some wide requests to receive
suboptimal resources from ec2cluster.

This change simplifies requirements. It now comprises a minmum
resource requirement and an integer width. A width of zero represents
a narrow request: the requirement's minimum represents the exact
resource requirement. Nonzero widths represent wide requests (even if
its width is 1). When we allocate instances in ec2cluster, we attempt
to expand requests to their width. The pool allocator is modified to
account for the new definition of width.

Along the way, we also update ec2cluster's status display.

TBA: To Be Audited.

Reviewers: pgopal

Differential Revision: https://phabricator.grailbio.com/D10435

fbshipit-source-id: 168c700