reflow/internal/status: pad winsize struct to accomodate for platform differences
Summary:
Darwin defined 4 extra bytes (two additional u16 fields) in its
winsize struct. This caused the ioctl call to corrupt some memory,
which in turn happened to somehow cause os.Stat(".") to fail, which
in turn would fail Reflow evaluation when given a relative path
argument, because the evaluator needs to resolve absolute paths.

The proper fix for this would be to fetch the true struct definition
using cgo, but we'd like to avoid this to preserve
cross-compileability to the greatest extent possible.

Reviewers: pgopal, razvanm

Reviewed By: razvanm

Differential Revision: https://phabricator.grailbio.com/D10982

fbshipit-source-id: 8bc27c4