grailbio/base/digest: add support for 512-bit hashes like SHA-512
Summary:
This commit adds support for 512-bit hashes like SHA-512 by increasing the
size of the array that stores the digest.

This commit fixes a bug related to encoding the hash in WriteDigest.
This commit fixes a bug related to decoding the hash in ReadDigest.

This commit fixes a couple of formatting errors in digest_test.go that
"go vet" complains about.

Test Plan:
Added test cases to digest_test.go that reproduce the failures while
trying to use SHA-512 digests for file archival. The tests pass after
the code changes.

Added test cases to reproduce bug in WriteDigest.

Ref: T5071

Reviewers: cnicolaou, marius, razvanm, O3 third-party, ysaito

Reviewed By: marius, O3 third-party, ysaito

Subscribers: ysaito

Differential Revision: https://phabricator.grailbio.com/D11018

fbshipit-source-id: 2e8c090