reflow/ec2cluster: config-overridable probe depth
Summary:
Ec2cluster probes EC2 (with a "dry run" request) to gauge available
spot capacity on a per-instance basis. The depth for this was hard
coded (at 20), which is more than the default instance limit for many
instance types in AWS. This change introduces a new config key for
ec2cluster, spotprobedepth, that allows a user to override the depth.
When set to 0 (also its default), capacity probing is disabled
entirely.

We maintain a depth of 20 for GRAIL's reflow distribution.

Reviewers: pgopal, razvanm

Reviewed By: razvanm

Differential Revision: https://phabricator.grailbio.com/D11292

fbshipit-source-id: 62d06a2