reflow/syntax: ensure that force resolution happens at most once
Summary:
Reflow's evaluator does not fully canonicalize chains of K's. Thus if
there's a K that returns another K, and the second K canonicalizes,
it can be applied twice. This is ok, except that maps (and modules,
but these are likely not problematic in practice) update an
underlying shared map on resolution. The code assumed it would be
called just once.

This change fixes this by making sure that resolution happens at most
one time. See the discussion at
https://github.com/grailbio/reflow/issues/28 for more details.

Reviewers: pgopal

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D11731

fbshipit-source-id: 37cce07