reflow: reflow.Collect, "reflow collect" tool
Summary:
Collect performs garbage collection of reflow caches, removing entries
that have not been accessed since a threshold date.  Begins by
traversing the cache building up livesets of cache keys and values.
Then removes cache entries not in the liveset from the repository and
assoc for the cache.  Adds a notion of a "threshold" time to a liveset
since new entries may have been added to the cache while the
collection is being performed.

Collect accepts a command line argument specifying the threshold date
to use for the live cutoff.

Reviewers: marius, pgopal

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D11551

fbshipit-source-id: ac379fe