reflow/pipelinerun: run in local mode and export local pool
Summary:
This change adds a new exclusive run command for pipeline runs.
The evaluator runs in a local executor. A local pool is exported
on a http port. If a v23 prefix is provided, the current invocation
will be announced on that prefix.

TODO: When the local run completes, everything is torn down. There is no
cooperative waiting for existing execs to complete. Not sure if we want
this.

Reviewers: marius, rck

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D12291

fbshipit-source-id: fce8d15