reflow/syntax: don't compute digests when an expression is immediate
Summary:
We needlessly computed digests for dependent values of immediate
expressions. This lead to, e.g., an "accidentally quadratic" map
(well, O(#lookups*#mapsize)) lookup since digesting has to fully
explore the underlying value.

When the expression that produces the map is not immediate, it's no
problem since the flow essentially memoizes the digest anyway. We
could also change the value representation to cache digests, or have
a shadow digest environment in the evaluator. However, I think that
this change should take care of the majority of expensive digest
computation.

Reviewers: pgopal, cmelton, sbagaria, joshnewman

Reviewed By: sbagaria

Differential Revision: https://phabricator.grailbio.com/D17325

fbshipit-source-id: 7af5cc2