refactor readme to make install more clear.
Summary:
Nearly everyone who is reading that should be downloading binaries, but the lede is buried and instead people smack their head against the wall with `go get`.

This moves the instructions for installing reflow for development purposes much lower in the page.
GitHub-Pull-Request: https://github.com/grailbio/reflow/pull/66

Reviewers: pgopal, joshnewman

Reviewed By: joshnewman

Differential Revision: https://phabricator.grailbio.com/D17693

fbshipit-source-id: f96e053