reflow/syntax: carefully label and promote flow types
Summary:
Reflow's types has a Flow "polymorphism": i.e., a bit in the type
describing whether a value of that type may be produced by a flow
(intern, exec, extern). This is used to statically check that, e.g.,
flow requirements are statically computable. In practice, this was
broken. This change fixes it by carefully promoting flow types and
testing them.

This will become increasingly useful in the future as we intend to
enforce flow requirements for some module parameters, etc., so
that we can guarantee that configuration (e.g., references) may
be statically computed, and does not depend on any external
comptuation.

Reviewers: smahadevan, pgopal, spolakh, sbagaria

Reviewed By: sbagaria

Differential Revision: https://phabricator.grailbio.com/D17996

fbshipit-source-id: 77c7dfa