Add a name field to the EC2Cluster configuration.
Summary:
This supports the ability to run multiple EC2 clusters
by implicitly creating different local state files
based on the cluster name.

In order to use this feature, one will have to modify
the config locally and pass it for subsequent reflow commands.

Fixes: https://github.com/grailbio/reflow/issues/65

Revised changes based on discussion in D18010

Test Plan:
Tested this with the following steps
  - reflow config -marshal > /tmp/config
  - modified /tmp/config with these changes into /tmp/config1:
> ec2clustername: "" -> ec2clustername: "custom1"
  - Now, the following commands will instantiate different EC2 instances.
```
  reflow -config=/tmp/config run <workflow.rf>
  reflow -config=/tmp/config1 run <workflow.rf>
```
  - Confirmed that different local state files exist.

Reviewers: marius, pgopal

Reviewed By: marius

Subscribers: marius

Differential Revision: https://phabricator.grailbio.com/D18043

fbshipit-source-id: 34ccafe