reflow/flow: allow submitting tasks via scheduler
Summary:
This change adds a new eval configuration, Scheduler. If set, the
evaluator constructs tasks and submits them asynchronously to the
provided scheduler instead of using the executor directly. The
scheduler is also responsible for data movement. The evaluator
assumes that the scheduler uses the cache repository for storing
objects. (And a repository must always be provided in this mode.)

Scheduler and Executor are mutually exclusive: they represent two
different modes of executing tasks. Once the scheduler matures, we
can start thinking about removing the executor altogether.

Reviewers: pgopal, smahadevan, sbagaria

Reviewed By: pgopal, sbagaria

Differential Revision: https://phabricator.grailbio.com/D19296

fbshipit-source-id: 859a0b0