reflow: support delayed loading of interns
Summary:
reflow: support delayed loading of interns

This change adds support to delay-load files and directories in Reflow.
Delay-loaded files are not loaded until they are used by an exec, at
which point they are pipelined with the actual execution.

Delay-loading happens in two phases:

1. File references are created in lieu of interns. These result in
   filesets that contain references (source+size+etag). These objects
   are passed around like regular files, but they do not represent
   interned files that have a Reflow ID.
2. Once they need to be used by an exec, the executor loads them
   directly, verifying their size and etag.

If a resolution fails, the code falls back to the regular intern+exec
If a path. load fails, evaluation fails.

The by-reference file objects that are passed around "just work" from the
point of view of caching: their cache key is derived from source+size+etag
instead of the file's SHA256.

With this support, we can now easily pass s3 URLs (plus etags)
directly into applications: their dependencies are captured correctly
by Reflow's cache mechanism. We can modify grailfile's URL schemes to
accept an "expected" etag, and then fail reads if that etag changes.
For extra assurance, we could also monitor the etag from Reflow
itself, to make sure that it does not change during execution.

Note that, in an effort to keep this change small(ish), it contains only
the minimum implementation required for delay-loading. There are a
number of follow-ups we should do:

1. After a reference file has been loaded once, we can now translate
   the remaining values to use the loaded file instead of re-loading
   from the reference. This could be done directly by rewriting values
   (see Fileset.Subst).
2. The complexity of S3 handling is getting out of hand. This should
   be simplified into a single client implementation that does everything
   we need to, including a central, unified way of resource/IO queueing.
   This mechanism should probably also handle other URL types (e.g.,
   localfile) in a uniform fashion.

Reviewers: smahadevan, sbagaria, pgopal

Reviewed By: sbagaria, pgopal

Differential Revision: https://phabricator.grailbio.com/D19439

fbshipit-source-id: ece8447