reflow/blob: introduce generalized blob store abstraction; use throughout
Summary:
This change excises S3 from Reflow, and encapsulates it within a
standard "blob store" abstraction. This has the effect of unifying S3
handling throughout Reflow, and also clean up and isolate the various
components that touch S3.

The new abstractions, blob.Store, and blob.Bucket, are general enough
to be used throughout Reflow, but also simple enough so that we could
easily provide, e.g., a GCS, or G3 implementations.

The new abstraction seamlessly handles region discovery and client
caching on a per bucket basis, so that we can treat S3 as uniform,
region-agnostic storage system.

Reviewers: pgopal, smahadevan

Reviewed By: pgopal

Differential Revision: https://phabricator.grailbio.com/D19582

fbshipit-source-id: fc86126