reflow: scalable evaluation
Summary:
This change introduces a number of efficiency improvements to
Reflow's flow evaluator. The most substantial (and intrusive) of
which is to walk only the minimal set of nodes required after each
node completion. Thus, after this change, we don't walk the entire
unfinished graph after each node completion, but instead walk only
the set of nodes that may change their state as the result of another
node's completion: each node is walked a constant number of times.

We also make some efficiency improvements to FlowVisitor.

Reviewers: pgopal, cmelton, sbagaria

Reviewed By: pgopal, sbagaria

Subscribers: spolakh

Maniphest Tasks: T10206

Differential Revision: https://phabricator.grailbio.com/D19993

fbshipit-source-id: c413c59