reflow/tool: resolver to canonicalize image names
Summary:
Images from a compiler session are resolved to their canonical
representations and passed on to EvalResult and then to EvalConfig
such that the flow can be canonicalized before cache lookup.

In the future, other remote resources can be similarly canonicalized.

For the time being, I am maintaining the original image names, and
computing an additional physical digest based on it.

An immediate consequence of this change is that when the same image is
specified in a different way (for example, using a different tag),
reflow will be able to reuse the cached values from a previous
execution.

This feature is not available for batch mode operations (`runbatch` and
`repair -batch`)

LINT: ONLY_NEW

Test Plan:
Unit tests and integration tests on harbormaster; inspecting debug logs
to verify that three cache keys are being generated when images with
tags are specified, and two cache keys when images with digests are
specified.

Reviewers: marius, pgopal

Reviewed By: marius

Subscribers: escott, cmelton

Maniphest Tasks: T10208

Differential Revision: https://phabricator.grailbio.com/D20390

fbshipit-source-id: 0b7a914