Add upper-bound to admission controller token growth
Summary:
Currently the admission controller's concurrency limit increases are unbounded
(as long as no errors are seen, that is).
Instead, allow clients to set an absolute max for the growth.

A single Acquire() request that asks for more tokens than the absolute max
will still be satisfied as long as no other tokens are used.
However, a successful Release() will not grow the token limit beyond that max.

Test Plan: Unit tests should cover most scenarios.

Reviewers: marius, pgopal

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D21298

fbshipit-source-id: 2c5a477