reflow/tool: Add a serve command which makes reflow binary run as a reflowlet.
Summary:
Add a serve command to reflow to make it run as a reflowlet.
Change the release reflow scripts to build reflow binary.
Change instance.go to reflect the new command-line
for running reflow as a reflowlet.

T14013

Test Plan: Tested manually by releasing and running a sample RF script.

Reviewers: marius, pgopal

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D22326

fbshipit-source-id: 5cef1b9