reflow/eval: fix squiggle bug
Summary:
In v := a ~> b, we weren't depending on b in the flow dependencies.
So if we ran this once and b changed (and is a flow), we end up
not evaluating the changed b.

Test Plan:
I am not super sure how I can write a test for this because of the caching effects.
I could possible write a unit test to check that the dependency exists unless you have
better ideas.

Reviewers: marius

Reviewed By: marius

Subscribers: cmelton, smahadevan

Differential Revision: https://phabricator.grailbio.com/D22607

fbshipit-source-id: cddf8b1