reflow/ec2cluster: fallback on version check if there's no embedded image
Summary:
The bazel rule reflow_binary does not do embedding.
If there's no embedded image, just fallback on version
check and successfully recognize a reflow server instance
that's up and running.

T14414

Test Plan:
Tested with and without this change with a sample program
as mentioned in the maniphest task and it works.

Reviewers: marius, pgopal, jcharumilind

Reviewed By: marius

Subscribers: sbagaria

Maniphest Tasks: T14414

Differential Revision: https://phabricator.grailbio.com/D23077

fbshipit-source-id: dbc2c71