reflow: Synchronously update state when adding instances
Summary:
A previous change () allowed for asynchronously updating
the pool whenever state changes happen.  This can cause
the addition of a new instance to not be recognized
immediately and we end up launching another instance.

This change makes update of the pool synchronous only
in the case of additions.

Test Plan:
Ran a sample reflow program multiple times
and ensured that we didn't start multiple instances.
(Before this change, that'll happen almost always).

Reviewers: marius, pgopal, jcharumilind

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D23089

fbshipit-source-id: 8f48fc6