reflow/eval: push down cheap operation timeouts, change default assoc concurrency
Summary:
Jobs with very large lookup concurrencies could have failed cache
lookups that were really just the result of queueing. This held true
both for the assoc lookup as well as file presence testing. In this
change, we push timeout decisions for "cheap" operations down to the
implementation, allowing for queueing to not be accounted for in
timeouts.

Given our large dynamoDB capacities, we also bump the default
concurrency limits for GRAIL's reflow config.

Reviewers: pgopal, smahadevan, jcharumilind

Reviewed By: smahadevan

Subscribers: cmelton

Differential Revision: https://phabricator.grailbio.com/D23495

fbshipit-source-id: 124daac