reflow/syntax: fix comparison for upcasted structs
Summary:
Structs may be upcast, and thus their values can contain more fields
than are present in their type. Deep comparison for structs is
implemented by converting the record value to a tuple: however, this
was done on a value level, not by type, and thus the resulting value
could contain more fields than was comparable. This caused two
issues: for two upcast records, we compared fields we should not be
comparing, thus comparison could fail when it should not; if one
record was upcast but another was not, this would cause an evaluator
panic since the tuple comparison code assumed that the tuples were
type-equal (which is a good assumptions--tuples cannot be upcast).

Fix this by constructing the tuples from the type, not the left
value.

Reviewers: pgopal, smahadevan, jcharumilind

Reviewed By: jcharumilind

Differential Revision: https://phabricator.grailbio.com/D23966

fbshipit-source-id: 95b7487