reflow/syntax: Reject non-terminal tail patterns
Summary: The current grammar for list tail patterns allows for tail patterns in the middle of the list, e.g. `[a, ..., b]`.  This is wrong and confusing.  This revision fixes this problem by only allowing tail patterns to come at the end of the list.

Test Plan: Unit tests; manual tests

Reviewers: marius, pgopal, smahadevan

Reviewed By: marius, pgopal

Differential Revision: https://phabricator.grailbio.com/D24761

fbshipit-source-id: 8ff17c2