reflow/cmd/releasereflow: update version file before building bootstrap
Summary:
This is a stopgap solution to be able to push an oss release.
We need to fix this issue with bootstrap binary expecting a version.

Also surface fatal and unavailable errors in debug log mode when we fail
to successfully provision an instance.

Test Plan: tried a release and ran a test on my personal aws account.

Reviewers: marius

Reviewed By: marius

Subscribers: smahadevan

Differential Revision: https://phabricator.grailbio.com/D26106

fbshipit-source-id: 1edde89