reflow/assoc: implement dynamodb batchget interface
Summary: The way reflow evaluator works today is that it does a bunch of cache lookups and then decides to evaluate some flows. Once those flows return, it does a bunch of new lookups and the process repeats. This bursty lookup pattern is best served with a batch interface so that we make the least number of aws API calls.

Test Plan: unit tests. will run end to end tests when I hook this up.

Reviewers: marius, jcharumilind, smahadevan

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D26385

fbshipit-source-id: 61af005