reflow: Make raising of FD limit work on OS X/Go 1.12
Summary:
This commit implements a workaround to an issue [[ https://github.com/golang/go/issues/30401 |  introduced in Go 1.12 ]].

The behavior of `Setrlimit` changed to no longer succeed when the requested limit exceeded the maximum.  (It used to succeed, clamping the value to the allowed maximum).

Test Plan: Manual testing

Reviewers: smahadevan, pgopal, marius

Reviewed By: marius

Maniphest Tasks: T17776

Differential Revision: https://phabricator.grailbio.com/D27211

fbshipit-source-id: 8be13f0