reflow/syntax: Implement variant expression digest computation
Summary:
The following Reflow program will cause a panic:
```lang=go
@requires(cpu := 1, mem := 128*MiB, disk := 128*MiB)
Main := delay(true) && #Foo == #Foo
```

This happens because this program triggers an attempt to compute a digest for the variant construction expression, which was not implemented.  This revision implements it.

Test Plan: Unit tests; manual testing

Reviewers: marius, smahadevan, pgopal

Reviewed By: marius

Maniphest Tasks: T17945

Differential Revision: https://phabricator.grailbio.com/D27382

fbshipit-source-id: 5f104dd