reflow/syntax: Make type error for unzip less confusing
Summary:
The error message when you call `unzip` with a non-tuple is confusing.  For example, consider this program:

```lang=go
Main := unzip(["a"])
```

This will fail with the message: `unzip expects a list of tuples, not string`.  This reads like `unzip` is expecting a list of a list of tuples, instead of a list of strings.  This revision makes the language consistently refer to the entire argument, so it will now say `unzip expects a list of tuple elements, not a list of string elements`.

Test Plan: Unit tests

Reviewers: pgopal, marius, joshnewman

Reviewed By: marius, joshnewman

Differential Revision: https://phabricator.grailbio.com/D30323

fbshipit-source-id: 2498da5