reflow/assertions: Make scheduler work with assertions
Summary:
Make scheduler work with assertions (regardless of policy).

Snapshots of files will now include assertions (which is needed to make
cache assertions work in scheduler mode).

When using scheduler (-sched) with no assertion policy
(ie, -assert never, which is also the default)
if an asserted object (eg: s3 blob file) does change,
the snapshotter will cause the changed object to be intern-ed.

Cache lookups of depending flow nodes will only succeed
if the propagation of assertions from dependent nodes is consistent.

More on the consistency check:
In scheduler mode, cached results are keyed by physical as well as logical
digests of every flow, but only the physical digests depend on the
snapshot (of say an S3 file).  When the snapshot changes, the physical
digest will not yield a cached result but the logical digest will.
However, the stored assertions of the cached result of a given flow
can now be inconsistent with the computed result of a dependent node.
The consistency check invalidates such a cached result.

Note: Before this change, reflow silently yields inconsistent results.

Depends on D29759

T19181

Test Plan: Same as described in D29759 with flags to use scheduler (-sched).

Reviewers: jcharumilind, pgopal, marius

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D30682

fbshipit-source-id: c7fa3de