reflow/flow: BottomUp mode should lookup the correct physical digest
Summary:
Ensure that flow.Argmap is defined when computing physical digest for an exec.

The physical digest depends on flow.Argmap which was otherwise being defined
only when we are ready to exec the flow (ie, when flow.ExecConfig() is called).
This means that when we lookup the cache, the physical digest key is different
from when we write it.

Test Plan: Unit tests

Reviewers: marius, pgopal, jcharumilind

Reviewed By: marius

Subscribers: sbanville

Maniphest Tasks: T23379

Differential Revision: https://phabricator.grailbio.com/D32529

fbshipit-source-id: 256691d