reflow/blob: Set ContentHash of S3 objects in metadata
Summary:
When reflow puts/copies objects in S3, it should set
the ContentHash (SHA256) in the object's metadata.

We also allow for the ContentHash to be set when
copying an S3 object, but only if the src object doesn't
already have it.

Note: This is an interim solution to address T24051 and we will revisit
this implementation when we refactor to make it cleaner.

Test Plan: Unit tests

Reviewers: marius, pgopal, jcharumilind

Reviewed By: marius, jcharumilind

Subscribers: spolakh

Maniphest Tasks: T25004

Differential Revision: https://phabricator.grailbio.com/D34917

fbshipit-source-id: a7da635