reflowlet, reflow/local: adjust OOM scores to prefer killing managed tasks
Summary:
This change ensures that the reflowlet (and its controlling Docker
invocation) runs with an OOM score adjustment of -1000, while tasks
managed by the Reflowlet are adjusted to +1000. The effect of this
should be that managed tasks are always first in line to be killed.
This in turn means that we should lose fewer tasks in OOM conditions;
that the reflowlet should remain usable; and that the user should
receive "killed by the OOM killer" error messages instead of
timeouts.

Reviewers: pgopal, smahadevan, jcharumilind, dnicolaou

Reviewed By: smahadevan, dnicolaou

Differential Revision: https://phabricator.grailbio.com/D35885

fbshipit-source-id: 62e058a